frames: clear strucutre and dependencies
